spostato creazione plot in plot.py
